Added problem solutions
Most of these are correct but some of them are not finished/do not work.
I think I have comments in these cases. I will go back later and fix
them when I have time. I also have a list of other problems I have
solved that are not in this group.